+++ src/Ledger/LedgerDetails/ExportSettingsModal/utils/getFiscalYearData.js
@@ -1,0 +1,3 @@
+import { FISCAL_YEARS_API } from '../../../../common/const';
+
+export const getFiscalYearData = (ky) => (fiscalYearId) => ky.get(`${FISCAL_YEARS_API}/${fiscalYearId}`).json();