+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Select filter should announce the number of Results in the Results List pane header. Refs UIF-371.
 * Display export action for ledger. Refs UIF-365.
 * User must select fiscal year and expense class settings for export. Refs UIF-366.
+* Start export budgets with confirmed settings. Refs UIF-367.
+* Create csv file with all budget data. Refs UIF-368.
 
 ## [3.1.1](https://github.com/folio-org/ui-finance/tree/v3.1.1) (2022-03-31)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v3.1.0...v3.1.1)